Move NavBar scroll animation to NavBar component file
